requirements.txt
amqp==5.0.9
asgiref==3.5.0
autoflake==1.4
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Pillow==9.1.0
Django==4.0.1
django-celery-results==2.2.0
django-filter==21.1
django-fsm==2.8.0
django-environ==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==4.8.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.3
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.26
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
redis==4.1.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3